Started to implement attack in CombatScene Added methods to caculate attack and resistance + Added Attack and Resistance to LivingThing, so started to repair all that is broken

+++ src/main/java/com/naheulback/nhlbck/classes/Weapon.java
@@ -13,7 +13,7 @@ public class Weapon extends Item {
 
     public double getQuality(){ return quality; }
 
-    public void decreaseQuality(int factor){
+    public void decreaseQuality(double factor){
         quality -= factor;
     }
